# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Apollo14 Air Quality API",
        "description": "API for air quality monitoring and photo submissions",
        "contact": {
            "email": "support@apollo14.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development server"
        }
    ],
    "paths": {
        "/air-quality/current": {
            "get": {
                "tags": [
                    "Air Quality"
                ],
                "summary": "Get current air quality for a location",
                "description": "Retrieves current air quality data including AQI, NO2, O3, and PM2.5 levels for the specified coordinates",
                "operationId": "f9bf07e3cdb9aab11be57c871f7e6c41",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitude coordinate",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "maximum": 90,
                            "minimum": -90,
                            "example": 40.7128
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Longitude coordinate",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "maximum": 180,
                            "minimum": -180,
                            "example": -74.006
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "aqi_value": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "no2_level": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 25.5
                                                },
                                                "o3_level": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 45.2
                                                },
                                                "pm25_level": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 12.8
                                                },
                                                "timestamp": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-15T10:30:00Z"
                                                },
                                                "data_sources": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "EPA",
                                                        "OpenAQ"
                                                    ]
                                                },
                                                "location": {
                                                    "properties": {
                                                        "latitude": {
                                                            "type": "number",
                                                            "example": 40.7128
                                                        },
                                                        "longitude": {
                                                            "type": "number",
                                                            "example": -74.006
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid location coordinates"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch air quality data"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Service unavailable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/air-quality/forecast": {
            "get": {
                "tags": [
                    "Air Quality"
                ],
                "summary": "Get 7-day air quality forecast",
                "description": "Retrieves 7-day air quality forecast data for the specified coordinates",
                "operationId": "8aada491f410cb97293a57881b9222d5",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitude coordinate",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "maximum": 90,
                            "minimum": -90,
                            "example": 40.7128
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Longitude coordinate",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "maximum": 180,
                            "minimum": -180,
                            "example": -74.006
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-01-15"
                                                    },
                                                    "aqi_value": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "no2_level": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 25.5
                                                    },
                                                    "o3_level": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 45.2
                                                    },
                                                    "pm25_level": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 12.8
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid location coordinates"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch forecast data"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Service unavailable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}